Problem Definition And Solution

DataDrivenDiffEq.DataDrivenProblem — Type

struct DataDrivenProblem{dType} <: DataDrivenDiffEq.AbstractDataDrivenProblem

The DataDrivenProblem defines a general estimation problem given measurements, inputs and (in the near future) observations. Two construction methods are available:

DiscreteDataDrivenProblem for time discrete systems

ContinousDataDrivenProblem for systems continouos in time

both are aliases for constructing a problem.

Fields

X

State measurements

t

Time measurements (optional)

DX

Differental state measurements (optional) or measurements of the next state

Y

Output measurements (optional; not used right now)

U

Input measurements (optional)

p

Parameters associated with the problem (optional)

is_discrete

(Time) discrete problem

Signatures

Example

X, DX, t = data...

# Define a discrete time problem
prob = DiscreteDataDrivenProblem(X)

# Define a continous time problem without explicit time points
prob = ContinuousDataDrivenProblem(X, DX)

# Define a continous time problem without explict derivatives
prob = ContinuousDataDrivenProblem(X, t)

# Define a discrete time problem with an input function as a function
input_signal(u,p,t) = t^2
prob = DiscreteDataDrivenProblem(X, t, input_signal)

DataDrivenDiffEq.DataDrivenSolution — Type

struct DataDrivenSolution{R<:DataDrivenDiffEq.AbstractBasis, S, P, A, O, M} <: DataDrivenDiffEq.AbstractDataDrivenSolution

The solution to a DataDrivenProblem derived via a certain algorithm. The solution is represented via an AbstractBasis, which makes it callable.

Fields

res

Result

retcode

Returncode

ps

Parameters used to derive the equations

alg

Algorithm used for solution

out

Original Output

inp

Original Input

metrics

Error metrics