Solutions

DataDrivenDiffEq.DataDrivenSolution — Type

struct DataDrivenSolution{L, A, O} <: DataDrivenDiffEq.AbstractDataDrivenSolution

The solution to a DataDrivenProblem derived via a certain algorithm. The solution is represented via an AbstractBasis, which makes it callable.

Fields

basis

The basis representation of the solution

parameters

Parameters of the solution

retcode

Returncode

alg

Algorithm

out

Original output of the solution algorithm

prob

Problem

l2_error

L₂ norm error

aic

AIC

rsquared

Coefficient of determinantion

Note

The L₂ norm error, AIC and coefficient of determinantion get only computed, if eval_expression is set to true or if the solution can be interpreted as a linear regression result.

Functions

DataDrivenDiffEq.result — Function

result(r)

Returns the Basis of the result.

DataDrivenDiffEq.get_problem — Function

get_problem(r)

Returns the original DataDrivenProblem.

DataDrivenDiffEq.metrics — Function

metrics(r)

Returns all applicable metrics of the solution.

ModelingToolkit.parameters — Function

parameters(r)

Returns the estimated parameters in form of an Vector.

DataDrivenDiffEq.parameter_map — Function

parameter_map(r)

Generate a mapping of the parameter values and symbolic representation useable to solve and ODESystem.

DataDrivenDiffEq.aic — Function

aic(r)

Returns the AIC of the result.

DataDrivenDiffEq.l2error — Function

l2error(r)

Returns the L₂ norm error of the result.

DataDrivenDiffEq.determination — Function

Returns the coefficient of determinantion of the result, if the result has been derived via a linear regression, e.g. sparse regression or koopman.

DataDrivenDiffEq.algorithm — Function

algorithm(r)

Returns the algorithm used to derive the solution.

DataDrivenDiffEq.output — Function

output(r)

Returns the original output of the algorithm.