Symbolic Regression

Using sparse regression limits the discovery to a generalized linear model where it is assumed that the nonlinear basis can capture the underlying function properly. Another approach is to use a general expression tree, which commonly encodes the function to discover as a binary tree where the nodes represent unary or binary operators acting on their children. DataDrivenDiffEq includes the following symbolic regression algorithms.

SymbolicRegression

Warning

This feature requires the explicit loading of SymbolicRegression.jl in addition to DataDrivenDiffEq. It will only be useable if loaded like:

using DataDrivenDiffEq
using SymbolicRegression

DataDrivenDiffEq provides an interface to SymbolicRegression.jl to solve a DataDrivenProblem:

using DataDrivenDiffEq
using LinearAlgebra
using Random
using SymbolicRegression

Random.seed!(1223)
# Generate a multivariate function for SymbolicRegression
X = rand(2,20)
f(x) = [sin(x[1]); exp(x[2])]
Y = hcat(map(f, eachcol(X))...)

# Define the options
opts = EQSearch([+, *, sin, exp], maxdepth = 1, progress = false, verbosity = 0)

# Define the problem
prob = DirectDataDrivenProblem(X, Y)

# Solve the problem
res = solve(prob, opts, numprocs = 0, multithreading = false)
sys = result(res)

Model ##Basis#81188 with 2 equations
States : x[1] x[2]
Independent variable: t
Equations
φ₁ = sin(x[1])
φ₂ = exp(x[2])

Where solve is used with EQSearch, which wraps Options provided by SymbolicRegression.jl. Additional keyworded arguments are max_iter = 10, which defines the number of iterations, weights which weight the measurements of the dependent variable (e.g. X, DX or Y depending on the DataDrivenProblem), numprocs which indicates the number of processes to use, procs for use with manually setup processes, multithreading = false for multithreading and runtests = true which performs initial testing on the environment to check for possible errors. It mimics the behaviour of EquationSearch.

Related Types

DataDrivenDiffEq.EQSearch — Type

struct EQSearch <: DataDrivenDiffEq.AbstractSymbolicRegression

Options for using SymbolicRegression.jl within the solve function. Automatically creates Options with the given specification. Sorts the operators stored in functions into unary and binary operators on conversion.

Fields

functions

Operators used for symbolic regression

kwargs

Additionally keyworded arguments passed to SymbolicRegression.Options

OccamNet

Warning

This feature requires the explicit loading of Flux.jl in addition to DataDrivenDiffEq. It will only be useable if loaded like:

using DataDrivenDiffEq
using Flux

As introduced in Interpretable Neuroevolutionary Models for Learning Non-Differentiable Functions and Programs , OccamNet is a special form of symbolic regression which uses a probabilistic approach to equation discovery by using a feedforward multilayer neural network. In contrast to normal architectures, each layer's weights reflect the probability of which inputs to use. Additionally a set of activation functions is used, instead of a single function. Similar to simulated annealing, a temperature is included to control the exploration of possible functions.

DataDrivenDiffEq offers two main interfaces to OccamNet: a Flux based API with Flux.train! and a solve(...) function.

Consider the following example, where we want to discover a vector valued function:

using DataDrivenDiffEq
using LinearAlgebra
using ModelingToolkit
using Flux
using Random

# Due to random values
Random.seed!(1223)

# Generate a multivariate dataset
X = rand(2,10)
f(x) = [sin(π*x[2]+x[1]); exp(x[2])]
Y = hcat(map(f, eachcol(X))...)

2×10 Matrix{Float64}:
 0.730729  0.999265  0.969019  0.625678  …  0.669184  0.95687  0.92499
 1.04471   1.29513   1.69627   1.74007      1.5878    1.36212  1.20932

Next, we define our network:

net = OccamNet(2, 2, 3, Function[sin, +, *, exp], skip = true, constants = Float64[π])

OccamNet(4, Constants 1, Parameters 0)

Where 2,2,3 refers to input and output dimension and the number of layers without the output layer. We also define that each layer uses the functions sin, +, *, exp as activations and uses a π as a constant, which get concatenated to the input data. Additionally, skip indicates the useage of skip connections, which allow the output of each layer to be passed onto the output layer directly.

To train the network over 100 epochs using ADAM, we type

Flux.train!(net, X, Y, ADAM(1e-2), 100, routes = 100, nbest = 3)

Under the hood, we select possible routes, routes, through the network based on the probability reflected by the ProbabilityLayer forming the network. From these we take the nbest candidates to train the parameters of the network, meaning increase the probability of those routes.

Lets have a look at some possible equations after the initial training. We can use rand to sample a route through the network, compute the output probability with probability and transform it into analytical equations by simply using ModelingToolkits variables as input. The call net(x, route) uses the route to compute just the element on this path.

@variables x[1:2]

for i in 1:10
  route = rand(net)
  prob = probability(net, route)
  eq = simplify.(net(x, route))
  print(eq , " with probability ",  prob, "\n")
end

Symbolics.Num[sin(exp(x[2])), exp(x[2]) + x[2]] with probability [0.02874208875253813, 0.001020356452889173]
Symbolics.Num[sin(x[1]*(3.141592653589793 + x[1])), exp(x[2])] with probability [2.4268035492071496e-5, 0.10004718259206825]
Symbolics.Num[exp(x[2]), exp(x[2])] with probability [0.010309792918739237, 0.011230891123571646]
Symbolics.Num[x[1], 3.141592653589793 + x[1]] with probability [0.05388524671461696, 0.0011426040901520863]
Symbolics.Num[-0.9125775986692777, exp(x[2])] with probability [0.006522697942050591, 0.02707702334938692]
Symbolics.Num[19333.689074365135, exp(x[2])] with probability [0.0006005975724607248, 0.10004718259206825]
Symbolics.Num[sin(exp(x[2])), x[1]] with probability [0.047892106785117454, 0.04565968696366679]
Symbolics.Num[sin(exp(x[2])), exp(x[2])] with probability [0.047892106785117454, 0.10004718259206825]
Symbolics.Num[1.2246467991473532e-16, exp(x[2])] with probability [0.017199511543924042, 0.02707702334938692]
Symbolics.Num[exp(x[2]^2), x[2]^2] with probability [0.0015392645555503143, 0.00936874746648553]

We see the networks proposals are not very certain. Hence, we will train for some more epochs and look at the output again.

Flux.train!(net, X, Y, ADAM(1e-2), 900, routes = 100, nbest = 3)

for i in 1:10
  route = rand(net)
  prob = probability(net, route)
  eq = simplify.(net(x, route))
  print(eq , " with probability ",  prob, "\n")
end

Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(exp(x[2])), exp(x[2])] with probability [0.002425719500856979, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]
Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [0.8965549834356896, 0.956500412755264]

The network is quite certain about the equation now, which is in fact our unknown mapping. To extract the solution with the highest probability, we set the temperature of the underlying distribution to a very low value. In the limit of t ↦ 0 we approach a Dirac distribution, hence extracting the most likely terms.

set_temp!(net, 0.01)
route = rand(net)
prob = probability(net, route)
eq = simplify.(net(x, route))
print(eq , " with probability ",  prob, "\n")

Symbolics.Num[sin(x[1] + 3.141592653589793x[2]), exp(x[2])] with probability [1.0, 1.0]

The same procedure is automated in the solve function. Using the same data, we wrap the algorithms information in the OccamSR struct and define a DataDrivenProblem:

# Define the problem
ddprob = DirectDataDrivenProblem(X, Y)
# Define the algorithm
sr_alg = OccamSR(functions = Function[sin, +, *, exp], skip = true, layers = 3, constants = [π])
# Solve the problem
res = solve(ddprob, sr_alg, ADAM(1e-2), max_iter = 1000, routes = 100, nbest = 3)

Explicit Result
Solution with 2 equations and 0 parameters.
Returncode: sucess
L2 Norm Error: 0.0
AICC: Inf

Within solve the network is generated using the information provided by the DataDrivenProblem in form of states, control and independent variables as well as the specified options, followed by training the network and extracting the equation with the highest probability by setting the temperature as above. After computing additional metrics, a DataDrivenSolution is returned where the equations are transformed into a Basis useable with ModelingToolkit.

The metrics can be accessed via:

metrics(res)

(Probability = 0.8425497124024455, Error = 0.0, AICC = Inf, Probabilities = [0.898724142309076, 0.9374953589626484], Errors = [0.0, 0.0], AICCs = [Inf, Inf])

and the resulting Basis by:

result(res)

Model ##Basis#81191 with 2 equations
States : x[1] x[2]
Independent variable: t
Equations
Differential(t)(x[1]) = sin(x[1] + 3.141592653589793x[2])
Differential(t)(x[2]) = exp(x[2])

Info

Right now, the resulting basis is not using parameters, but raw numerical values.

Related Types

DataDrivenDiffEq.OccamNet — Type

mutable struct OccamNet{F, C, P} <: DataDrivenDiffEq.AbstractOccam

Defines an OccamNet which learns symbolic expressions from data using a probabilistic approach. See Interpretable Neuroevolutionary Models for Learning Non-Differentiable Functions and Programs for more details.

It get constructed via:

net = OccamNet(inp::Int, outp::Int, layers::Int, f::Vector{Function}, t::Real = 1.0; constants = typeof(t)[], parameters::Int = 0, skip::Bool = false, init_w = ones, init_p = Flux.glorot_uniform)

inp describes the size of the input domain, outp the size of the output domain, layers the number of layers (including the input layer and excluding the linear output layer) and f the functions to be used. Optional is the temperature t which is set to 1.0 at the beginning.

Keyworded arguments are constants, a vector of constants like π, ℯ which can concatenated to the input, the number of trainable parameters and if skip connections should be used. The constructors to the weights and parameters can be passed in via init_w and init_p.

OccamNet is callable with and without a specific route, which can be sampled from the networks weights via rand(net).

Fields

c

The Chain representing the network

constants

Additional constants added to the input which are not trainable.

parameters

Additional parameters added to the input which are trainable.

DataDrivenDiffEq.OccamSR — Type

struct OccamSR{F, C, T} <: DataDrivenDiffEq.AbstractSymbolicRegression

Options for using OccamNet within the solve function. Automatically creates a network with the given specification.

Fields

functions

Functions used within the network

constants

Constants added to the input

layers

Number of layers

parameters

Number of parameters

skip

Activate skip connections

DataDrivenDiffEq.ProbabilityLayer — Type

mutable struct ProbabilityLayer{F, W, T, A} <: DataDrivenDiffEq.AbstractProbabilityLayer

Defines a basic ProbabilityLayer in which the parameters act as probabilities via the softmax function for an array of functions.

The layer is callable either via layer(x), using all weights to form the output or by layer(x, route) where route is the result of rand(layer) which samples the function arguments from the underlying distribution.

Fields

op

Nonlinear functions forming the basis of the layer

weight

Weights

t

Temperature controlling the shape of the distribution

arieties

Arieties of the functions

skip

Skip connection

Related Functions

DataDrivenDiffEq.set_temp! — Function

set_temp!(p, t)

Set the temperature of the ProbabilityLayer or OccamNet.

DataDrivenDiffEq.probability — Function

probability(o, route)

Returns the probability of the result of the OccamNet using the specific route.

DataDrivenDiffEq.logprobability — Function

logprobability(o, route)

Returns the logprobability of the result of the OccamNet using the specific route.

DataDrivenDiffEq.probabilities — Function

probabilities(p)

Return the probability associated with the ProbabilityLayer or OccamNet by applying softmax on the weights.

DataDrivenDiffEq.logprobabilities — Function

logprobabilities(p)

Return the logprobability associated with the ProbabilityLayer or OccamNet by applying logsoftmax on the weights.