Koopman Operators

Missing docstring.

Missing docstring for LinearKoopman. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NonlinearKoopman. Check Documenter's build log for details.

Functions

DataDrivenDiffEq.operator — Function

operator(k)

Return the approximation of the discrete Koopman operator stored in k.

DataDrivenDiffEq.generator — Function

generator(k)

Return the approximation of the continuous Koopman generator stored in k.

Missing docstring.

Missing docstring for inputmap. Check Documenter's build log for details.

DataDrivenDiffEq.outputmap — Function

outputmap(k)

Return the array C, mapping the Koopman space back onto the state space.

DataDrivenDiffEq.update! — Function

update!(k, X, Y; p, t, U, threshold)

Update the Koopman k given new data X and Y. The operator is updated in place if the L2 error of the prediction exceeds the threshold.

p and t are the parameters of the basis and the vector of timepoints, if necessary.

DataDrivenDiffEq.is_discrete — Function

is_discrete(k)

Returns if the AbstractKoopmanOperator k is discrete in time.

DataDrivenDiffEq.is_continuous — Function

is_continuous(k)

Returns if the AbstractKoopmanOperator k is continuous in time.

DataDrivenDiffEq.updatable — Function

updatable(k)

Returns true if the AbstractKoopmanOperator is updatable.

Missing docstring.

Missing docstring for isstable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eigen. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eigvals. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eigvecs. Check Documenter's build log for details.

DataDrivenDiffEq.modes — Function

modes(k)

Return the eigenvectors of a continuous AbstractKoopmanOperator.

DataDrivenDiffEq.frequencies — Function

frequencies(k)

Return the eigenvalues of a continuous AbstractKoopmanOperator.

Missing docstring.

Missing docstring for reduce_basis. Check Documenter's build log for details.