Problem Definition And Solution

As can be seen from the introduction examples, DataDrivenDiffEq.jl tries to structurize the workflow in a similar fashion to other SciML packages by defining a DataDrivenProblem, dispatching on the solve command to return a DataDrivenSolution.

A problem in the sense of identification, estimation or inference is defined by the data describing it. This data contains at least measurements of the states X, which would be sufficient to describe a DiscreteDataDrivenProblem with unit time steps similar to the first example on dynamic mode decomposition. Of course we can extend this to include time points t, consecutive measurements X̃ at the next time point or control signals U or a function describing those u(x,p,t). Additionally, any parameters p known a priori can be included in the problem. In practice, this looks like

problem = DiscreteDataDrivenProblem(X)
problem = DiscreteDataDrivenProblem(X, t)
problem = DiscreteDataDrivenProblem(X, t, X̃)
problem = DiscreteDataDrivenProblem(X, t, X̃, U = U)
problem = DiscreteDataDrivenProblem(X, t, X̃, U = U, p = p)
problem = DiscreteDataDrivenProblem(X, t, X̃, U = (x,p,t)->u(x,p,t))

Similarly, a ContinuousDataDrivenProblem would need at least measurements and time-derivatives (X and DX) or measurements, time information and a way to derive the time derivatives(X, t and a Collocation method). Again, this can be extended by including a control input as measurements or a function and possible parameters.

problem = ContinuousDataDrivenProblem(X, t, InterpolationMethod())
problem = ContinuousDataDrivenProblem(X, DX = DX)
problem = ContinuousDataDrivenProblem(X, t, DX = DX)
problem = ContinuousDataDrivenProblem(X, t, DX = DX, U = U)
problem = ContinuousDataDrivenProblem(X, t, DX = DX, U = U, p = p)
problem = ContinuousDataDrivenProblem(X, t, DX = DX, U = (x,p,t)->u(x,p,t))

Next up, we choose a method to solve the DataDrivenProblem. Depending on the input arguments and the type of problem, the function will return a result derived via Koopman or Sparse Optimization methods. Different options can be provided as well as a Basis used for lifting the measurements, to control different options like rounding, normalization or the progressbar depending on the inference method. Possible options are provided below.

# Use a Koopman based inference
res = solve(problem, DMDSVD(), kwargs...)
# Use a sparse identification
res = solve(problem, basis, STLQS(), kwargs...)

The DataDrivenSolution res contains a result which is the infered system and a Basis, metrics which is a NamedTuple containing different metrics like the L2 error of the infered system with the provided data and the AICC. These can be accessed via

# The infered system
system = result(res)
# The metrics
m = metrics(res)
m.Sparsity # No. of active terms / nonzero coefficients
m.Error # L2 Error of all data
m.Errors # Individual error of the different data rows
m.AICC # AICC
m.AICCs # ....

Since the infered system is a parametrized equation, the corresponding parameters can be accessed and returned via

# Vector
ps = parameters(res)
# Parameter map
ps = parameter_map(res)

Optional Arguments

Koopman based algorithms can be called without a Basis, resulting in dynamic mode decomposition like methods, or with a basis for extened dynamic mode decomposition :

res = solve(problem, DMDSVD(), kwargs...)
res = solve(problem, basis, DMDSVD(), kwargs...)

If control signals are present, they get processed according to this paper for dynamic mode decomposition and as described here for extended dynamic mode decomposition assuming a linear relationship on the operator.

Possible keyworded arguments include

B a linear mapping known a priori which maps the control signals onto the lifted states

digits controls the digits / rounding used for deriving the system equations (digits = 1 would round 10.02 to 10.0)

operator_only returns a NamedTuple containing the operator, input and output mapping and matrices used for updating the operator as described here

SINDy based algorithms can be called like :

res = solve(problem, basis, STLQS(), kwargs...)
res = solve(problem, basis, ImplicitOptimizer(), kwargs...)

Where control signals are included in the candidate basis.

For implicit optimizers, additionally an Vector{Num} of variables correspondng to $u_t$ for implicitly defined equations $f(u_t, u, p, t) = 0$ can be passed in

res = solve(problem, basis, ImplicitOptimizer(), implicits, kwargs...)

This indicates that we do not have coupling between two implicitly defined variables if not explicitly given. To elaborate this complicated sentence, consider the following:

basis = Basis([x, y, z], [x,y,z])
res = solve(problem, basis, ImplicitOptimizer())

Would allow solutions of the form x = y + z.

basis = Basis([x, y, z], [x,y,z])
res = solve(problem, basis, ImplicitOptimizer(), [x,y])

Would exclude solutions of the form x = y + z since we separate x and y. However, defining

basis = Basis([x, y, z, x*y], [x,y,z])
res = solve(problem, basis, ImplicitOptimizer(), [x,y])

Would allow solutions of the form x = y*x + z or y = y*x + z to occur while suppressing x = y + z. This is due to the reason that y*x includes both x and y, so the function will get included in the evaluation.

Possible keyworded arguments include

normalize normalizes the data matrix $\\Theta$ such that each row ( correspondng to candidate functions) has a 2-norm of 1.0

denoise applies optimal shrinking to the matrix $\\Theta$ to remove the influence of noise

maxiter Maximum iterations of the used optimizer

round rounds according to the currently used threshold of the optimizer

Types

DataDrivenDiffEq.DataDrivenProblem — Type

struct DataDrivenProblem{dType} <: DataDrivenDiffEq.AbstractDataDrivenProblem

The DataDrivenProblem defines a general estimation problem given measurements, inputs and (in the near future) observations. Two construction methods are available:

DiscreteDataDrivenProblem for time discrete systems

ContinousDataDrivenProblem for systems continouos in time

both are aliases for constructing a problem.

Fields

X

State measurements

t

Time measurements (optional)

DX

Differental state measurements (optional) or measurements of the next state

Y

Output measurements (optional; not used right now)

U

Input measurements (optional)

p

Parameters associated with the problem (optional)

is_discrete

(Time) discrete problem

Signatures

Example

X, DX, t = data...

# Define a discrete time problem
prob = DiscreteDataDrivenProblem(X)

# Define a continous time problem without explicit time points
prob = ContinuousDataDrivenProblem(X, DX)

# Define a continous time problem without explict derivatives
prob = ContinuousDataDrivenProblem(X, t)

# Define a discrete time problem with an input function as a function
input_signal(u,p,t) = t^2
prob = DiscreteDataDrivenProblem(X, t, input_signal)

DataDrivenDiffEq.DataDrivenSolution — Type

struct DataDrivenSolution{R<:DataDrivenDiffEq.AbstractBasis, S, P, A, O, M} <: DataDrivenDiffEq.AbstractDataDrivenSolution

The solution to a DataDrivenProblem derived via a certain algorithm. The solution is represented via an AbstractBasis, which makes it callable.

Fields

res

Result

retcode

Returncode

ps

Parameters used to derive the equations

alg

Algorithm used for solution

out

Original Output

inp

Original Input

metrics

Error metrics

Functions

DataDrivenDiffEq.result — Function

result(r)

Returns the result of in form of an AbstractBasis.

ModelingToolkit.parameters — Function

parameters(r)

Returns the estimated parameters in form of an Vector.

DataDrivenDiffEq.parameter_map — Function

parameter_map(r)

Generate an mapping of the parameter values and symbolic representation useable to solve and ODESystem.

DataDrivenDiffEq.metrics — Function

metrics(r)

Returns the metrics of the result in form of a NamedTuple.

DataDrivenDiffEq.output — Function

output(r)

Returns the original output of the algorithm, e.g. an AbstractArray of coefficients for sparse regression.

DataDrivenDiffEq.algorithm — Function

algorithm(r)

Returns the algorithm used to derive the solution.

DataDrivenDiffEq.inputs — Function

inputs(r)

Returns the original inputs, most commonly the DataDrivenProblem and the Basis used to derive the solution.