Solutions

DataDrivenDiffEq.DataDrivenSolution — Type

struct DataDrivenSolution{R<:Union{Nothing, DataDrivenDiffEq.AbstractBasis}, S, P, A, O, M} <: DataDrivenDiffEq.AbstractDataDrivenSolution

The solution to a DataDrivenProblem derived via a certain algorithm. The solution is represented via an AbstractBasis, which makes it callable.

Fields

res

Result

retcode

Returncode

ps

Parameters used to derive the equations

alg

Algorithm used for solution

out

Original Output

inp

Original Input

metrics

Error metrics

Functions

DataDrivenDiffEq.result — Function

result(r)

Returns the result of in form of an AbstractBasis.

ModelingToolkit.parameters — Function

parameters(r)

Returns the estimated parameters in form of an Vector.

DataDrivenDiffEq.parameter_map — Function

parameter_map(r)

Generate a mapping of the parameter values and symbolic representation useable to solve and ODESystem.

DataDrivenDiffEq.metrics — Function

metrics(r)

Returns the metrics of the result in form of a NamedTuple.

DataDrivenDiffEq.output — Function

output(r)

Returns the original output of the algorithm, e.g. an AbstractArray of coefficients for sparse regression.

DataDrivenDiffEq.algorithm — Function

algorithm(r)

Returns the algorithm used to derive the solution.

DataDrivenDiffEq.inputs — Function

inputs(r)

Returns the original inputs, most commonly the DataDrivenProblem and the Basis used to derive the solution.