Solutions

DataDrivenDiffEq.DataDrivenSolution — Type

struct DataDrivenSolution{T} <: DataDrivenDiffEq.AbstractDataDrivenSolution

The solution to a DataDrivenProblem derived via a certain algorithm. The solution is represented via an Basis, which makes it callable.

Fields

basis: The basis representation of the solution

retcode: Returncode

alg: Algorithm

out: Original output of the solution algorithm

prob: Problem

residuals: Residual sum of squares

dof: Degrees of freedom

internal_problem: Internal problem

API

DataDrivenDiffEq.get_problem — Function

get_problem(r)

Returns the original DataDrivenProblem.

DataDrivenDiffEq.get_basis — Function

get_basis(r)

Returns the recovered Basis.

DataDrivenDiffEq.get_algorithm — Function

get_algorithm(r)

Returns the algorithm used to derive the solution.

DataDrivenDiffEq.get_results — Function

get_results(r)

Returns the original output of the algorithm.

DataDrivenDiffEq.is_converged — Function

is_converged(r)

Assert the result of the DataDrivenSolution and returns true if successful, false otherwise.

Additionally, DataDrivenDiffEq.jl extends the following methods for a DataDrivenSolution.

StatsAPI.dof — Function

dof(sol)

Returns the degrees of freedom of the DataDrivenSolution.

StatsAPI.rss — Function

rss(sol)

Returns the residual sum of squares of the DataDrivenSolution.

StatsAPI.nobs — Function

nobs(sol)

Returns the number of observations of the DataDrivenSolution.

StatsAPI.loglikelihood — Function

loglikelihood(sol)

Returns the loglikelihood of the DataDrivenSolution assuming a normal distributed error.

StatsAPI.nullloglikelihood — Function

nullloglikelihood(sol)

Return the nullloglikelihood of the DataDrivenSolution. This corresponds to a model only fitted with an intercept and a normal distributed error.

StatsAPI.r2 — Function

r2(sol)

Return the coefficient of determinantion of the DataDrivenSolution.

Note

Only implements CoxSnell based on the loglikelihood and nullloglikelihood.

StatsBase.summarystats — Function

summarystats(sol)

Returns the summarystats for each row of the error for the DataDrivenSolution.