Optimizer

DataDrivenDiffEq comes with some implementations for sparse regression included. All of these are stored inside the DataDrivenDiffEq.Optimize package and extend the AbstractOptimizer, if an explicit optimization is needed or the AbstractSubspaceOptimizer for an implicit problem (where the solution is within the nullspace).

Functions

Missing docstring.

Missing docstring for STRRidge. Check Documenter's build log for details.

DataDrivenDiffEq.Optimize.ADMM — Type

mutable struct ADMM{T, R} <: DataDrivenDiffEq.Optimize.AbstractOptimizer{T}

ADMM is an implementation of Lasso using the alternating direction methods of multipliers and loosely based on this implementation. It solves the following problem

\[\min_{x} \frac{1}{2} \| Ax-b\|_2 + \lambda \|x\|_1\]

#Fields

λ

Sparsity threshold

ρ

Augmented Lagrangian parameter

Example

opt = ADMM()
opt = ADMM(1e-1, 2.0)

DataDrivenDiffEq.Optimize.SR3 — Type

mutable struct SR3{T, V, P<:DataDrivenDiffEq.Optimize.AbstractProximalOperator} <: DataDrivenDiffEq.Optimize.AbstractOptimizer{T}

SR3 is an optimizer framework introduced by Zheng et. al., 2018 and used within Champion et. al., 2019. SR3 contains a sparsification parameter λ, a relaxation ν. It solves the following problem

\[\min_{x, w} \frac{1}{2} \| Ax-b\|_2 + \lambda R(w) + \frac{\nu}{2}\|x-w\|_2\]

Where R is a proximal operator and the result is given by w. #Fields

λ

Sparsity threshold

ν

Relaxation parameter

R

Proximal operator

Example

opt = SR3()
opt = SR3(1e-2)
opt = SR3(1e-3, 1.0)
opt = SR3(1e-3, 1.0, SoftThreshold())

Note

Opposed to the original formulation, we use ν as a relaxation parameter, as given in Champion et. al., 2019. In the standard case of hard thresholding the sparsity is interpreted as λ = threshold^2 / 2, otherwise λ = threshold.

DataDrivenDiffEq.Optimize.ADM — Type

mutable struct ADM{T} <: DataDrivenDiffEq.Optimize.AbstractSubspaceOptimizer{T}

Optimizer for finding a sparse basis vector in a subspace based on this paper. It solves the following problem

\[\min_{x} \|x\|_0 ~s.t.~Ax= 0\]

#Fields

λ

Sparsity threshold

Example

ADM()
ADM(λ = 0.1)

Note

While useable for implicit problems, a better choice in general is given by the ImplicitOptimizer which tends to be more robust.

Implementing New Optimizer

Similarly to Algorithms for Estimation, the extension of optimizers is more or less straightforward. Suppose you want to define a new optimizer MyOpt, which should solve $A~X = Y$ for a sparse $X$.

mutable struct MyOpt <: DataDrivenDiffEq.Optimize.AbstractOptimizer
    threshold
end

To use MyOpt within SINDy, an init! function has to be implemented.

function init!(X::AbstractArray, o::MyOpt, A::AbstractArray, Y::AbstractArray)
    X .=  A \ Y
end

To perform thresholding - and use maybe for searching the right threshold - a setter and getter is required:

set_threshold!(opt::MyOpt, threshold) = opt.threshold .= threshold

get_threshold(opt::MyOpt) = opt.threshold

And, at last, the method which fits the data and returns the iterations needed:

function fit!(X::AbstractArray, A::AbstractArray, Y::AbstractArray, opt::MyOpt; maxiter, convergence_error)
    # Compute awesome stuff here
    return iterations
end