Optimizer

DataDrivenDiffEq comes with some implementations for sparse regression included. All of these are stored inside the DataDrivenDiffEq.Optimize package and extend the AbstractOptimizer, if an explicit optimization is needed or the AbstractSubspaceOptimizer for an implicit problem (where the solution is within the nullspace).

Functions

DataDrivenDiffEq.Optimize.STLSQ — Type

mutable struct STLSQ{T} <: DataDrivenDiffEq.Optimize.AbstractOptimizer

STLQS is taken from the original paper on SINDY and implements a sequentially thresholded least squares iteration. λ is the threshold of the iteration. It is based upon this matlab implementation.

It solves the following problem

\[\min_{x} \frac{1}{2} \| Ax-b\|_2 + \lambda \|x\|_2\]

#Fields

λ

Sparsity threshold

Example

opt = STLQS()
opt = STLQS(1e-1)

Note

This was formally STRRidge and has been renamed.

DataDrivenDiffEq.Optimize.ADMM — Type

mutable struct ADMM{T} <: DataDrivenDiffEq.Optimize.AbstractOptimizer

ADMM is an implementation of Lasso using the alternating direction methods of multipliers and loosely based on this implementation.

It solves the following problem

\[\min_{x} \frac{1}{2} \| Ax-b\|_2 + \lambda \|x\|_1\]

#Fields

λ

Sparsity threshold

ρ

Augmented Lagrangian parameter

Example

opt = ADMM()
opt = ADMM(1e-1, 2.0)

DataDrivenDiffEq.Optimize.SR3 — Type

mutable struct SR3{T, P} <: DataDrivenDiffEq.Optimize.AbstractOptimizer

SR3 is an optimizer framework introduced by Zheng et. al., 2018 and used within Champion et. al., 2019. SR3 contains a sparsification parameter λ, a relaxation ν.

It solves the following problem

\[\min_{x, w} \frac{1}{2} \| Ax-b\|_2 + \lambda R(w) + \frac{\nu}{2}\|x-w\|_2\]

Where R is a proximal operator and the result is given by w.

#Fields

λ

Sparsity threshold

ν

Relaxation parameter

R

Proximal operator

Example

opt = SR3()
opt = SR3(1e-2)
opt = SR3(1e-3, 1.0)
opt = SR3(1e-3, 1.0, SoftThreshold())

Note

Opposed to the original formulation, we use ν as a relaxation parameter, as given in Champion et. al., 2019. In the standard case of hard thresholding the sparsity is interpreted as λ = threshold^2 / 2, otherwise λ = threshold.

DataDrivenDiffEq.Optimize.ADM — Type

mutable struct ADM{U} <: DataDrivenDiffEq.Optimize.AbstractSubspaceOptimizer

Optimizer for finding a sparse basis vector in a subspace based on this paper.

It solves the following problem

\[\min_{x} \|x\|_0 ~s.t.~Ax= 0\]

#Fields

λ

Sparsity threshold

Example

ADM()
ADM(λ = 0.1)

Proximal Operators

DataDrivenDiffEq.Optimize.SoftThreshold — Type

struct SoftThreshold <: DataDrivenDiffEq.Optimize.AbstractProximalOperator

Proximal operator which implements the soft thresholding operator.

sign(x) * max(abs(x) - λ, 0)

See by Zheng et. al., 2018.

DataDrivenDiffEq.Optimize.HardThreshold — Type

struct HardThreshold <: DataDrivenDiffEq.Optimize.AbstractProximalOperator

Proximal operator which implements the hard thresholding operator.

abs(x) > sqrt(2*λ) ? x : 0

See by Zheng et. al., 2018.

DataDrivenDiffEq.Optimize.ClippedAbsoluteDeviation — Type

struct ClippedAbsoluteDeviation{T} <: DataDrivenDiffEq.Optimize.AbstractProximalOperator

Proximal operator which implements the (smoothly) clipped absolute deviation operator.

abs(x) > ρ ? x : sign(x) * max(abs(x) - λ, 0)

Where ρ = 5λ per default.

#Fields

ρ

Upper threshold

Example

opt = ClippedAbsoluteDeviation()
opt = ClippedAbsoluteDeviation(1e-1)

See by Zheng et. al., 2018.

Implementing New Optimizer

Similarly to Algorithms for Estimation, the extension of optimizers is more or less straightforward. Suppose you want to define a new optimizer MyOpt, which should solve $A~X = Y$ for a sparse $X$.

mutable struct MyOpt <: DataDrivenDiffEq.Optimize.AbstractOptimizer
    threshold
end

To use MyOpt within SINDy, an init! function has to be implemented.

function init!(X::AbstractArray, o::MyOpt, A::AbstractArray, Y::AbstractArray)
    X .=  A \ Y
end

To perform thresholding - and use maybe for searching the right threshold - a setter and getter is required:

set_threshold!(opt::MyOpt, threshold) = opt.threshold .= threshold

get_threshold(opt::MyOpt) = opt.threshold

And, at last, the method which fits the data and returns the iterations needed:

function fit!(X::AbstractArray, A::AbstractArray, Y::AbstractArray, opt::MyOpt; maxiter, convergence_error)
    # Compute awesome stuff here
    return iterations
end