Collocation

DataDrivenDiffEq.InterpolationMethod — Type

A wrapper for the interpolation methods of DataInterpolations.jl.

Wraps the methods in such a way that they are callable as f(u,t) to create and return an interpolation of u over t. The first argument of the constructor always defines the interpolation method, all following arguments will be used in the interpolation.

Example

# Create the wrapper struct
itp_method = InterpolationMethod(QuadraticSpline)
# Create a callable interpolation
itp = itp_method(u,t)
# Return u[2]
itp(t[2])

DataDrivenDiffEq.collocate_data — Function

collocate_data(data, tpoints)
collocate_data(data, tpoints, kernel)

Unified interface for collocation techniques. The input can either be a CollocationKernel (see list below) or a wrapped InterpolationMethod from DataInterpolations.jl.

Computes a non-parametrically smoothed estimate of u' and u given the data, where each column is a snapshot of the timeseries at tpoints[i].

Examples

u′,u = collocate_data(data,tpoints,kernel=SigmoidKernel())
u′,u = collocate_data(data,tpoints,tpoints_sample,interp,args...)
u′,u = collocate_data(data,tpoints,interp)

Collocation Kernels

See this paper for more information.

EpanechnikovKernel

UniformKernel

TriangularKernel

QuarticKernel

TriweightKernel

TricubeKernel

GaussianKernel

CosineKernel

LogisticKernel

SigmoidKernel

SilvermanKernel

Interpolation Methods

See DataInterpolations.jl for more information.

ConstantInterpolation

LinearInterpolation

QuadraticInterpolation

LagrangeInterpolation

QuadraticSpline

CubicSpline

BSplineInterpolation

BSplineApprox

Curvefit

Utilities

DataDrivenDiffEq.AIC — Function

AIC(k, X, Y; likelihood)

Computes the Akaike Information Criterion (AIC) given the free parameters k for the data X and its estimate Y of the model. likelihood can be any function of X and Y.

DataDrivenDiffEq.AICC — Function

AICC(k, X, Y; likelihood)

Computes the Akaike Information Criterion compensated for finite samples (AICC) given the free parameters k for the data X and its estimate Y of the model. likelihood can be any function of X and Y.

DataDrivenDiffEq.BIC — Function

BIC(k, X, Y; likelihood)

Computes Bayes Information Criterion (BIC) given the free parameters k for the data X and its estimate Y of the model. likelihood can be any function of X and Y.

DataDrivenDiffEq.optimal_shrinkage — Function

optimal_shrinkage(X)

Compute a feature reduced version of the data array X via thresholding the singular values by computing the optimal threshold for singular values.

DataDrivenDiffEq.burst_sampling — Function

(

burst_sampling(x, samplesize, bursts)

)

Randomly selects n bursts of data with size samplesize from the data X.

Randomly selects n bursts of data with size samplesize from the data X and Y.

Randomly selects n bursts of data within a time window period from the data X. The time information has to be provided in t.

DataDrivenDiffEq.subsample — Function

subsample(x, frequency)

Returns the subsampled X with only every n-th entry.

Returns the subsampled X with a a minimum period of dt between two data points. t provides the time information.